Refactor repeated focusMap component
