+++ apps/antalmanac/src/components/Calendar/CourseCalendarEvent.tsx
@@ -3,9 +3,10 @@ import { Theme, withStyles } from '@material-ui/core/styles';
 import { ClassNameMap, Styles } from '@material-ui/core/styles/withStyles';
 import { Delete, Search } from '@material-ui/icons';
 import { WebsocSectionFinalExam } from '@packages/antalmanac-types';
-import { useEffect, useRef, useCallback } from 'react';
+import { useEffect, useRef } from 'react';
 import { Event } from 'react-big-calendar';
-import { Link } from 'react-router-dom';
+
+import { MapLink } from '../buttons/MapLink';
 
 import { deleteCourse, deleteCustomEvent } from '$actions/AppStoreActions';
 import CustomEventDialog from '$components/Calendar/Toolbar/CustomEventDialog/';
@@ -15,8 +16,7 @@ import buildingCatalogue from '$lib/buildingCatalogue';
 import { clickToCopy, useQuickSearchForClasses } from '$lib/helpers';
 import locationIds from '$lib/location_ids';
 import AppStore from '$stores/AppStore';
-import { useTimeFormatStore, useThemeStore } from '$stores/SettingsStore';
-import { useTabStore } from '$stores/TabStore';
+import { useTimeFormatStore } from '$stores/SettingsStore';
 import { formatTimes } from '$stores/calendarizeHelpers';
 
 const styles: Styles<Theme, object> = {
@@ -164,15 +164,9 @@ const CourseCalendarEvent = (props: CourseCalendarEventProps) => {
         };
     }, []);
 
-    const { setActiveTab } = useTabStore();
     const quickSearch = useQuickSearchForClasses();
 
     const { isMilitaryTime } = useTimeFormatStore();
-    const isDark = useThemeStore((store) => store.isDark);
-
-    const focusMap = useCallback(() => {
-        setActiveTab(2);
-    }, [setActiveTab]);
 
     const { classes, selectedEvent } = props;
 
@@ -262,14 +256,10 @@ const CourseCalendarEvent = (props: CourseCalendarEventProps) => {
                             <td className={`${classes.multiline} ${classes.rightCells}`}>
                                 {locations.map((location) => (
                                     <div key={`${sectionCode} @ ${location.building} ${location.room}`}>
-                                        <Link
-                                            className={classes.clickableLocation}
-                                            to={`/map?location=${locationIds[location.building] ?? 0}`}
-                                            onClick={focusMap}
-                                            color={isDark ? '#1cbeff' : 'blue'}
-                                        >
-                                            {location.building} {location.room}
-                                        </Link>
+                                        <MapLink
+                                            buildingId={locationIds[location.building] ?? '0'}
+                                            room={`${location.building} ${location.room}`}
+                                        />
                                     </div>
                                 ))}
                             </td>
@@ -302,13 +292,7 @@ const CourseCalendarEvent = (props: CourseCalendarEventProps) => {
                 {building && (
                     <div className={classes.table}>
                         Location:&nbsp;
-                        <Link
-                            className={classes.clickableLocation}
-                            to={`/map?location=${building ?? 0}`}
-                            onClick={focusMap}
-                        >
-                            {buildingCatalogue[+building]?.name ?? ''}
-                        </Link>
+                        <MapLink buildingId={+building} room={buildingCatalogue[+building]?.name ?? ''} />
                     </div>
                 )}
                 <div className={classes.buttonBar}>
